+++ haproxy-minio/haproxy-minio.d/local/share/cook/bin/configure-haproxy-1.sh
@@ -11,6 +11,10 @@ set -o pipefail
 
 export PATH=/usr/local/bin:$PATH
 
+# make directories
+mkdir -p /usr/local/etc/haproxy
+mkdir -p /var/run/haproxy
+
 SCRIPT=$(readlink -f "$0")
 TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
 
@@ -18,14 +22,14 @@ TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
 # safe(r) separator for sed
 sep=$'\001'
 
-# copy in custom nginx and set IP to ip address of pot image
-< "$TEMPLATEPATH/nginx3.conf.in" \
+# copy in 1 server haproxy.cfg
+< "$TEMPLATEPATH/haproxy-1.cfg.in" \
+  sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
   sed "s${sep}%%domain%%${sep}$DOMAIN${sep}g" | \
   sed "s${sep}%%serverone%%${sep}$SERVERONE${sep}g" | \
-  sed "s${sep}%%servertwo%%${sep}$SERVERTWO${sep}g" | \
-  sed "s${sep}%%serverthree%%${sep}$SERVERTHREE${sep}g" | \
-  sed "s${sep}%%bucket%%${sep}$BUCKET${sep}g"  \
-  > /usr/local/etc/nginx/nginx.conf
+  sed "s${sep}%%serveroneport%%${sep}$SERVERONEPORT${sep}g" \
+  > /usr/local/etc/haproxy/haproxy.conf
 
-# enable nginx
-service nginx enable || true
+# enable haproxy
+sysrc haproxy_config="/usr/local/etc/haproxy/haproxy.conf"
+service haproxy enable || true