+++ pixelfed/pixelfed.d/local/share/cook/bin/configure-cron.sh
@@ -1,0 +1,15 @@
+#!/bin/sh
+
+# shellcheck disable=SC1091
+if [ -e /root/.env.cook ]; then
+    . /root/.env.cook
+fi
+
+set -e
+# shellcheck disable=SC3040
+set -o pipefail
+
+export PATH=/usr/local/bin:$PATH
+
+# setup cronjob
+echo "*	*	*	*	*	root	sudo -u www /usr/local/bin/php /usr/local/www/pixelfed/artisan --no-ansi schedule:run > /dev/null" >> /etc/crontab